Deprecate `@FetchAll/One` parameter pack use
We maintain a ton of overloads to support them, but they are not very
ergonomic. We provide a far better tool for packing values together, and
that's the `@Selection` macro, so let's steer folks that direction.
